Move CT specific code back into CT
