+++ src/main/java/com/hotels/hcommon/hive/metastore/client/closeable/CloseableMetaStoreClientFactory.java
@@ -66,6 +66,4 @@ public final class CloseableMetaStoreClientFactory implements MetaStoreClientFac
         compatibility);
     return (CloseableMetaStoreClient) Proxy.newProxyInstance(classLoader, interfaces, handler);
   }
-
-
 }
